+++ src/pages/home/offers/FlightOffer.tsx
@@ -19,10 +19,11 @@ import {
   reactionLogic,
   selectedReaction,
 } from "../../../common/offers/reactions";
-import OfferFooterImage from "./OfferFooterImage";
 import LocalOfferIcon from "@mui/icons-material/LocalOffer";
 import dayjs from "dayjs";
 import { offerAvailability } from "../../../common/offers/functions";
+import ReserveOnline from "../../../common/ReserveOnline";
+import OfferFooterImage from "../../../common/OfferFooterImage";
 
 const FlightOffer = () => {
   const { get } = flightOffer();
@@ -230,6 +231,12 @@ const FlightOffer = () => {
               description: value.description,
               footerImage: (
                 <OfferFooterImage
+                  reserveBtn={
+                    <ReserveOnline
+                      id={value.id}
+                      availability={offerAvailability(value)}
+                    />
+                  }
                   price={value.price}
                   startDate={value.startDate}
                   endDate={value.endDate}
