+++ src/pages/home/offers/HotelOffer.tsx
@@ -19,10 +19,11 @@ import {
   reactionLogic,
   selectedReaction,
 } from "../../../common/offers/reactions";
-import OfferFooterImage from "./OfferFooterImage";
 import LocalOfferIcon from "@mui/icons-material/LocalOffer";
 import dayjs from "dayjs";
 import { offerAvailability } from "../../../common/offers/functions";
+import ReserveOnline from "../../../common/ReserveOnline";
+import OfferFooterImage from "../../../common/OfferFooterImage";
 
 const HotelOffer = () => {
   const { get } = hotelOffer();
@@ -219,6 +220,12 @@ const HotelOffer = () => {
               description: value.description,
               footerImage: (
                 <OfferFooterImage
+                  reserveBtn={
+                    <ReserveOnline
+                      id={value.id}
+                      availability={offerAvailability(value)}
+                    />
+                  }
                   price={value.price}
                   startDate={value.startDate}
                   endDate={value.endDate}
